requirements.txt
# Interfaces
gradio==5.24.0
fastapi==0.115.12
viser

# Computation
numpy==1.23.5
torch==2.5.1 --index-url https://download.pytorch.org/whl/cu124
torchvision==0.20.1 --index-url https://download.pytorch.org/whl/cu124
torchaudio==2.5.1 --index-url https://download.pytorch.org/whl/cu124
onnxruntime==1.15.1
# onnxruntime-gpu==1.17.1
onnxoptimizer==0.3.13
timm==0.1.12
einops
# flash_attn
roma # 3D rotation
splines

# Data Manipulation
pandas==2.2.3
pydantic==2.11.3
supervision==0.25.1
datasets==3.5.0

# Data Processing
opencv-python==4.7.0.72
pymatting
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0

# AI Models
accelerate==1.5.*
# accelerate==0.33.0
bitsandbytes==0.45.*
transformers==4.39.1
safetensors==0.5.*
pytorch-lightning==2.2.1
git+https://github.com/facebookresearch/segment-anything.git
# SAM-2 @ git+https://github.com/facebookresearch/segment-anything-2.git
# clip @ git+https://github.com/openai/CLIP.git
ultralytics
mediapipe
facexlib
geffnet==0.9.0
segmentation-models-pytorch
peft==0.15.*

# API
# SpeechRecognition==3.10.0
# flask_cloudflared==0.0.14
# sse-starlette==1.6.5
# tiktoken

# Others
numba==0.57.0
pooch==1.8.1
markdown
jsonschema==4.21.1
# av==8.0.3
# https://pyav.org/docs/develop/overview/installation.html#build-on-windows
pims==0.5
pyliblzfse
plyfile
yourdfpy
robot_descriptions
